Restore number-key tab switching in transfer mode
Transfer mode lost direct number-key tab switching when #1065 dropped
the dispatch_direct_tab_input call there, so 1-9/0 only reached the
"Use Tab to switch pane" warning. Move the helper to input_common and
reuse it from both browsing and transfer so tabs are reachable with
number keys in either mode.
