+++ docs/keybindings.ja.md
@@ -76,6 +76,7 @@ Transfer は `enter dir`、`. hidden`、`Tab switch-pane`、`p/Esc close`、`q q
 | キー | 動作 |
 | --- | ------ |
 | `Tab` / `Shift+Tab` | 反対側ペインへフォーカスを切り替え（転送方向を設定） |
+| `1`〜`0` | 数字キーで対応するタブへ切り替える（1が最初のタブ、0が10番目のタブ） |
 | `Esc` | 選択を解除、未選択時は通常モードに戻る |
 | `j` / `↓` | フォーカス中ペインで下へ移動 |
 | `k` / `↑` | フォーカス中ペインで上へ移動 |
@@ -101,7 +102,7 @@ Transfer は `enter dir`、`. hidden`、`Tab switch-pane`、`p/Esc close`、`q q
 | `p` | 通常モードに戻る |
 | `q` | アプリを終了 |
 
-`Tab` がペイン切替に使われるため、ブラウザタブ操作（新規/切替/閉じる）は転送モードのコマンドパレット（`:`）から利用します。
+ブラウザタブ操作は転送モードのコマンドパレット（`:`）から利用できます。タブの切替は数字キー（`1`〜`0`）で直接行え、`Tab` はペイン切替に使います。
 
 ---
