+++ docs/release-checklist.md
@@ -171,6 +171,7 @@
 - [ ] p キーまたは「Toggle transfer mode」で転送モードに切り替わること
 - [ ] 2ペインレイアウトと転送ヘッダー（方向・パス・件数）が表示されること
 - [ ] Tab キーで左右のペインフォーカスが切り替わり、方向表示も更新されること
+- [ ] 1-9, 0 キーで対応するタブに切り替わること
 - [ ] 各ペインで独立してナビゲーションできること
 - [ ] c キーでフォーカスペインの項目を反対ペインにコピーできること
 - [ ] m キーでフォーカスペインの項目を反対ペインに移動できること
